+++ tiled/client/register.py
@@ -340,11 +340,23 @@ async def register_single_item(
 
 # Matches filename with (optional) prefix characters followed by digits \d
 # and then the file extension .tif or .tiff.
-TIFF_SEQUENCE_STEM_PATTERN = re.compile(r"^(.*?)(\d+)\.(?:tif|tiff)$")
-TIFF_SEQUENCE_EMPTY_NAME_ROOT = "_unnamed"
+IMG_SEQUENCE_STEM_PATTERNS = {
+    ".tif": re.compile(r"^(.*?)(\d+)\.(?:tif|tiff)$"),
+    ".tiff": re.compile(r"^(.*?)(\d+)\.(?:tif|tiff)$"),
+    ".jpg": re.compile(r"^(.*?)(\d+)\.(?:jpg|jpeg)$"),
+    ".jpeg": re.compile(r"^(.*?)(\d+)\.(?:jpg|jpeg)$"),
+}
+IMG_SEQUENCE_EMPTY_NAME_ROOT = "_unnamed"
+
+IMG_SEQUENCE_MIMETYPES = {
+    ".tif": "multipart/related;type=image/tiff",
+    ".tiff": "multipart/related;type=image/tiff",
+    ".jpg": "multipart/related;type=image/jpeg",
+    ".jpeg": "multipart/related;type=image/jpeg",
+}
 
 
-async def group_tiff_sequences(
+async def group_image_sequences(
     node,
     path,
     files,
@@ -365,26 +377,35 @@ async def group_tiff_sequences(
     unhandled_files = []
     sequences = collections.defaultdict(list)
     for file in files:
-        if file.is_file():
-            match = TIFF_SEQUENCE_STEM_PATTERN.match(file.name)
+        file_ext = Path(file).suffixes[-1] if len(Path(file).suffixes) > 0 else None
+        if file.is_file() and file_ext and file_ext in IMG_SEQUENCE_STEM_PATTERNS:
+            match = IMG_SEQUENCE_STEM_PATTERNS[file_ext].match(file.name)
             if match:
                 sequence_name, _sequence_number = match.groups()
                 if sequence_name == "":
-                    sequence_name = TIFF_SEQUENCE_EMPTY_NAME_ROOT
+                    sequence_name = IMG_SEQUENCE_EMPTY_NAME_ROOT
                 sequences[sequence_name].append(file)
                 continue
         unhandled_files.append(file)
     for name, sequence in sequences.items():
-        await register_tiff_sequence(node, name, sorted(sequence), settings)
+        await register_image_sequence(node, name, sorted(sequence), settings)
     return unhandled_files, unhandled_directories
 
 
-TIFF_SEQ_MIMETYPE = "multipart/related;type=image/tiff"
-
-
-async def register_tiff_sequence(node, name, sequence, settings):
-    logger.info("    Grouped %d TIFFs into a sequence '%s'", len(sequence), name)
-    adapter_class = settings.adapters_by_mimetype[TIFF_SEQ_MIMETYPE]
+async def register_image_sequence(node, name, sequence, settings):
+    suffixes = Path(sequence[0]).suffixes
+    file_ext = suffixes[-1] if len(suffixes) > 0 else None
+    if file_ext and file_ext in IMG_SEQUENCE_MIMETYPES:
+        mimetype = IMG_SEQUENCE_MIMETYPES[file_ext]
+    else:
+        raise RuntimeError(f"Cannot register image sequence {name}.")
+    logger.info(
+        "    Grouped %d %s images into a sequence '%s'",
+        len(sequence),
+        mimetype.split("/")[-1],
+        name,
+    )
+    adapter_class = settings.adapters_by_mimetype[mimetype]
     key = settings.key_from_filename(name)
     try:
         adapter = adapter_class([ensure_uri(filepath) for filepath in sequence])
@@ -400,7 +421,7 @@ async def register_tiff_sequence(node, name, sequence, settings):
         data_sources=[
             DataSource(
                 structure_family=adapter.structure_family,
-                mimetype=TIFF_SEQ_MIMETYPE,
+                mimetype=mimetype,
                 structure=dict_or_none(adapter.structure()),
                 parameters={},
                 management=Management.external,
@@ -435,7 +456,7 @@ async def skip_all(
     return [], directories
 
 
-DEFAULT_WALKERS = [group_tiff_sequences, one_node_per_item]
+DEFAULT_WALKERS = [group_image_sequences, one_node_per_item]
 
 
 async def watch(